+++ client/src/components/Input/NewConversationMenu/EndpointItem.jsx
@@ -36,7 +36,7 @@ export default function ModelItem({ endpoint, value, isSelected }) {
         value={value}
         className={cn(
           'group dark:font-semibold dark:text-gray-100 dark:hover:bg-gray-800',
-          isSelected && 'dark:bg-gray-900'
+          isSelected && 'dark:bg-gray-900 bg-gray-50 active'
         )}
       >
         {icon}
