+++ e2e/specs/popup.spec.js
@@ -1,0 +1,24 @@
+import { expect, test } from '@playwright/test';
+
+test.describe('Endpoints Presets suite', () => {
+  let myBrowser;
+
+  test.beforeEach(async ({ browser }) => {
+    myBrowser = await browser.newContext({
+      storageState: './auth.json',
+    });
+  });
+
+  test('Endpoints Suite', async () => {
+    const page = await myBrowser.newPage();
+    await page.goto('http://localhost:3080/');
+    await page.getByRole('button', { name: 'New Topic' }).click();
+
+    const endpointItem = await page.getByRole('menuitemradio', { name: 'OpenAI' })
+    await endpointItem.click();
+
+    await page.getByRole('button', { name: 'New Topic' }).click();
+    // Check if the active class is set on the selected endpoint
+    expect(await endpointItem.getAttribute('class')).toContain('active');
+  });
+});
